requirements.txt
discord.py==1.7.3
humanize==4.3.0
py-dactyl==1.1.4
python-dotenv==0.20.0
PyYAML==6.0
